requirements.txt
pyspark==3.2.1
pandas
boto3
python-dotenv
